# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VisiteCRM API - SaaS CRM for tourism agencies
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List trips
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripListResponse"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripBody"
      responses:
        "201":
          description: Created trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Not found
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTripBody"
      responses:
        "200":
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Delete a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /trips/{id}/seat-map:
    get:
      operationId: getTripSeatMap
      tags: [trips]
      summary: Get trip seat map with occupation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Seat map
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeatMap"

  # ─── RESERVATIONS ─────────────────────────────────────────────────
...
  /users/me/sync:
    post:
      operationId: syncMe
      tags: [users]
      summary: Sync Clerk user into DB after login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncUserBody"
      responses:
        "200":
          description: Synced user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (sellers/team)
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user (invite seller)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    # ─── DASHBOARD ─────────────────────────────────────────────────
...
    Trip:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        destination:
          type: string
        destinationCity:
          type: string
        destinationState:
          type: string
        type:
          type: string
        category:
          type: string
        departureDate:
          type: string
        returnDate:
          type: ["string", "null"]
        totalCapacity:
          type: integer
        availableSeats:
          type: integer
        reservedSeats:
          type: integer
        confirmedSeats:
          type: integer
        priceAdult:
          type: number
        priceChild:
          type: ["number", "null"]
        priceSenior:
          type: ["number", "null"]
        inclusions:
          type: array
          items:
            type: string
        exclusions:
          type: array
          items:
            type: string
        coverImage:
          type: ["string", "null"]
        gallery:
          type: array
          items:
            type: string
        status:
          type: string
        isPublic:
          type: boolean
        isFeatured:
          type: boolean
        vehiclePlate:
          type: ["string", "null"]
        vehicleType:
          type: ["string", "null"]
        driverName:
          type: ["string", "null"]
        seatLayout:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - slug
        - destination
        - destinationCity
        - destinationState
        - type
        - category
        - departureDate
        - totalCapacity
        - availableSeats
        - reservedSeats
        - confirmedSeats
        - priceAdult
        - inclusions
        - exclusions
        - gallery
        - status
        - isPublic
        - isFeatured
        - createdAt
        - updatedAt
    TripListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Trip"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    CreateTripBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        destination:
          type: string
        destinationCity:
          type: string
        destinationState:
          type: string
        type:
          type: string
        category:
          type: string
        departureDate:
          type: string
        returnDate:
          type: ["string", "null"]
        totalCapacity:
          type: integer
        priceAdult:
          type: number
        priceChild:
          type: ["number", "null"]
        priceSenior:
          type: ["number", "null"]
        inclusions:
          type: array
          items:
            type: string
        exclusions:
          type: array
          items:
            type: string
        coverImage:
          type: ["string", "null"]
        seatLayout:
          type: string
        vehiclePlate:
          type: ["string", "null"]
        vehicleType:
          type: ["string", "null"]
        driverName:
          type: ["string", "null"]
      required:
        - name
        - destination
        - destinationCity
        - destinationState
        - type
        - category
        - departureDate
        - totalCapacity
        - priceAdult
    UpdateTripBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        isPublic:
          type: ["boolean", "null"]
        isFeatured:
          type: ["boolean", "null"]
        departureDate:
          type: ["string", "null"]
        returnDate:
          type: ["string", "null"]
        priceAdult:
          type: ["number", "null"]
        priceChild:
          type: ["number", "null"]
        priceSenior:
          type: ["number", "null"]
        totalCapacity:
          type: ["integer", "null"]
        coverImage:
          type: ["string", "null"]
        inclusions:
          type: ["array", "null"]
          items:
            type: string
        exclusions:
          type: ["array", "null"]
          items:
            type: string
        vehiclePlate:
          type: ["string", "null"]
        vehicleType:
          type: ["string", "null"]
        driverName:
          type: ["string", "null"]
    SeatMap:
      type: object
      properties:
        tripId:
          type: string
        layout:
          type: string
        totalSeats:
          type: integer
        seats:
          type: array
          items:
            $ref: "#/components/schemas/Seat"
      required: [tripId, layout, totalSeats, seats]
    Seat:
      type: object
      properties:
        number:
          type: string
        row:
          type: integer
        col:
          type: integer
        status:
          type: string
        passengerName:
          type: ["string", "null"]
        reservationId:
          type: ["string", "null"]
      required: [number, row, col, status]

    # ─── RESERVATION ───────────────────────────────────────────────
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        tenantId:
          type: ["string", "null"]
        referralCode:
          type: string
        referralBalance:
          type: number
        createdAt:
          type: string
      required: [id, clerkId, name, email, role, isActive, referralCode, referralBalance, createdAt]
    SyncUserBody:
      type: object
      properties:
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
      required: [clerkId, name, email]
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
      required: [name, email, role]
    UpdateUserBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]