# lib/api-zod/src/generated/types/seat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VisiteCRM API - SaaS CRM for tourism agencies
 * OpenAPI spec version: 0.1.0
 */

export interface Seat {
  number: string;
  row: number;
  col: number;
  status: string;
  /** @nullable */
  passengerName?: string | null;
  /** @nullable */
  reservationId?: string | null;
}